Fix imports after moving getBaseUrl utility method to a common file
