+++ src/commands/sarif/renderer.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk'
 import upath from 'upath'
 
-import {getBaseUrl} from '../junit/utils'
+import {getBaseUrl} from '../../helpers/app'
 
 import {Payload} from './interfaces'
